Changed type of Pj.m_Cd from float to int (it's expressed in physics ticks)

+++ Common/src/PjCommon.h
@@ -26,10 +26,10 @@ namespace rmkl {
 		std::map<int, PjState> m_History;
 
 	private:
-		float m_Cd;
+		int m_Cd;
 
 	private:
-		static const float s_CdDuration;
+		static const int s_CdDuration;
 	};
 
 }
